requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-docx==1.1.0
google-cloud-documentai==2.20.0
openai==1.3.0
werkzeug==3.0.1
gunicorn==21.2.0